Minor build fixes.
I made these two tiny changes to SeaBIOS when I was playing with the
HEAD version.

The Makefile change prevents the test command from failing if CC
contains a space.

The coreboot change prevents a warning on our gcc 4.5.2 in certain
circumstances when sizeof seems to be unsigned long int instead of
unsigned int.

They're both unintrusive, shouldn't have potential for breaking
anything.

+++ src/coreboot.c
@@ -320,7 +320,7 @@ ulzma(u8 *dst, u32 maxlen, const u8 *src, u32 srclen)
     u8 scratch[15980];
     int need = (LzmaGetNumProbs(&state.Properties) * sizeof(CProb));
     if (need > sizeof(scratch)) {
-        dprintf(1, "LzmaDecode need %d have %d\n", need, sizeof(scratch));
+        dprintf(1, "LzmaDecode need %d have %d\n", need, (unsigned int)sizeof(scratch));
         return -1;
     }
     state.Probs = (CProb *)scratch;
